docs: test custom docs for buttons
