+++ .storybook/main.js
@@ -5,7 +5,10 @@ module.exports = {
     '../projects/canopy/src/lib/alert/alert.stories.ts',
     '../projects/canopy/src/lib/brand-icon/brand-icons.stories.ts',
     '../projects/canopy/src/lib/breadcrumb/breadcrumb.stories.ts',
-    '../projects/canopy/src/lib/button/button.stories.ts',
+    '../projects/canopy/src/lib/button/docs/usage.stories.mdx',
+    '../projects/canopy/src/lib/button/docs/*.primary.stories.ts',
+    '../projects/canopy/src/lib/button/docs/*.secondary.stories.ts',
+    '../projects/canopy/src/lib/button/docs/*.icon.stories.ts',
     '../projects/canopy/src/lib/button/button-group/button-group.stories.ts',
     '../projects/canopy/src/lib/card/card.stories.ts',
     '../projects/canopy/src/lib/carousel/carousel.stories.ts',
