--- projects/canopy/src/lib/button/button.notes.mdx
@@ -1,42 +1,0 @@
-import { ArgsTable, Story, Canvas, PRIMARY_STORY } from '@storybook/addon-docs';
-
-# Button
-
-Provides styles and behaviours for Canopy buttons. `lg-button` is a directive which can be applied to any `<button>` or `<a>` element.
-
-Import the button module in your module:
-~~~js
-@NgModule({
-  ...
-  imports: [ ..., LgButtonModule ],
-});
-~~~
-
-<div class="lg-container">
-  <div class="lg-row">
-    <div class="lg-col-xs-12 lg-col-md-4">
-      Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
-    </div>
-    <div class="lg-col-xs-12 lg-col-md-4">
-      Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
-    </div>
-    <div class="lg-col-xs-12 lg-col-md-4">
-      Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
-    </div>
-  </div>
-</div>
-
-<br></br>
-
-## An example story
-<Story id="components-button--icon-only" />
-
-## Another story with the canvas
-<Canvas columns={2}  withSource="open" withToolbar>
-  <Story id="components-button--primary-dark" />
-</Canvas>
-
-## Args table
-Do we want this here or should we keep it only in the Canvas tab?
-
-<ArgsTable story={PRIMARY_STORY} />
