+++ projects/canopy/src/lib/button/docs/button.secondary.notes.mdx
@@ -1,0 +1,50 @@
+import {Canvas, Story, Source} from '@storybook/addon-docs';
+
+# Secondary button
+
+## Dark variant
+
+<Canvas withSource="open">
+  <Story id="components-button-previews--secondary-dark"/>
+</Canvas>
+
+## Light variant
+
+<Source id="components-button-previews--secondary-light"></Source>
+
+This button variant can be used on the below backgrounds:
+
+<Canvas columns={2}>
+  <div style={{
+    backgroundColor: '#0076d6',
+    color: 'white',
+    padding: '20px',
+    borderRadius: '4px',
+  }}>
+    <Story id="components-button-previews--secondary-light"/>
+  </div>
+  <div style={{
+    backgroundColor: '#333',
+    color: 'white',
+    padding: '20px',
+    borderRadius: '4px',
+  }}>
+    <Story id="components-button-previews--secondary-light"/>
+  </div>
+  <div style={{
+    backgroundColor: '#028844',
+    color: 'white',
+    padding: '20px',
+    borderRadius: '4px',
+  }}>
+    <Story id="components-button-previews--secondary-light"/>
+  </div>
+  <div style={{
+    backgroundColor: '#005380',
+    color: 'white',
+    padding: '20px',
+    borderRadius: '4px',
+  }}>
+    <Story id="components-button-previews--secondary-light"/>
+  </div>
+</Canvas>
