+++ projects/canopy/src/lib/button/docs/usage.notes.ts
@@ -1,0 +1,3 @@
+export const notes = `
+This is the usage
+`;
